Check if Loupe is set up when interacting with documents (#52)

+++ src/Internal/Index/Indexer.php
@@ -92,6 +92,10 @@ class Indexer
      */
     public function deleteDocuments(array $ids): self
     {
+        if ($this->engine->getIndexInfo()->needsSetup()) {
+            return $this;
+        }
+
         $this->engine->getConnection()
             ->executeStatement(
                 sprintf('DELETE FROM %s WHERE user_id IN(:ids)', IndexInfo::TABLE_NAME_DOCUMENTS),
